# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cars API V1",
        "description": "API для управления автомобилями, их комплектациями, опциями и ценами.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/configurations": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Получить список комплектаций",
                "description": "Возвращает список всех комплектаций, включая связанные опции и цены.",
                "operationId": "getConfigurations",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Configuration"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Создать новую комплектацию",
                "description": "Создает новую комплектацию автомобиля и возвращает созданную запись.",
                "operationId": "storeConfiguration",
                "requestBody": {
                    "description": "Данные для создания комплектации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Комплектация создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configuration"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/configurations/{configuration}": {
            "get": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Получить комплектацию по ID",
                "description": "Возвращает данные комплектации по заданному ID.",
                "operationId": "getConfigurationById",
                "parameters": [
                    {
                        "name": "configuration",
                        "in": "path",
                        "description": "ID комплектации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные комплектации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configuration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комплектация не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Обновить комплектацию",
                "description": "Обновляет данные комплектации по заданному ID.",
                "operationId": "updateConfiguration",
                "parameters": [
                    {
                        "name": "configuration",
                        "in": "path",
                        "description": "ID комплектации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления комплектации",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ConfigurationInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Комплектация обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Configuration"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Комплектация не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Configurations"
                ],
                "summary": "Удалить комплектацию",
                "description": "Удаляет комплектацию по заданному ID.",
                "operationId": "deleteConfiguration",
                "parameters": [
                    {
                        "name": "configuration",
                        "in": "path",
                        "description": "ID комплектации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Комплектация удалена"
                    },
                    "404": {
                        "description": "Комплектация не найдена"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Configuration": {
                "required": [
                    "id",
                    "car_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Comfort"
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Climate Control"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-19T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-19T12:34:56Z"
                    }
                },
                "type": "object"
            },
            "ConfigurationInput": {
                "required": [
                    "car_id",
                    "name"
                ],
                "properties": {
                    "car_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Comfort"
                    },
                    "option_ids": {
                        "description": "Массив ID опций",
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}